Develop (#2)
* init project added all entities controllers etc
added jwt authentication added code to add first user in database on first launch

* refactor fixed bad relation where address was linked to users instead of patients WARNING please drop docker container if exists before launch

+++ src/main/java/com/b3al/med/medi_nfo/patient/PatientResource.java
@@ -1,6 +1,8 @@
 package com.b3al.med.medi_nfo.patient;
 
 import com.b3al.med.medi_nfo.user.UserRole;
+import com.b3al.med.medi_nfo.util.ReferencedException;
+import com.b3al.med.medi_nfo.util.ReferencedWarning;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
@@ -91,6 +93,10 @@ public class PatientResource {
     @PreAuthorize("hasAuthority('" + UserRole.Fields.ADMIN + "')")
     @ApiResponse(responseCode = "204")
     public ResponseEntity<Void> deletePatient(@PathVariable(name = "id") final Long id) {
+        final ReferencedWarning referencedWarning = patientService.getReferencedWarning(id);
+        if (referencedWarning != null) {
+            throw new ReferencedException(referencedWarning);
+        }
         patientService.delete(id);
         return ResponseEntity.noContent().build();
     }
